[RENAME] hook 폴더 재구조
- hook 폴더 내 auth, user, post로 나눔
- hook 폴더 가시성을 위해 분리함
